+++ packages/vsce/CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the "cics-extension-for-zowe" extension will be documente
 ## Recent Changes
 
 - Enhancement: Updated highlight information for CICS Resources: JVM Server, TS Queue, Bundle, Bundle Part, JVM Endpoint, Library Dataset. [#573](https://github.com/zowe/cics-for-zowe-client/pull/573)
+- Enhancement: Added Compare command to view two CICS resources side by side. [#560](https://github.com/zowe/cics-for-zowe-client/issues/560)
 
 ## `3.18.0`